feat(nativebridgeif): Move nativebridgeif src; Update drivers. - Move iob_nativebridgeif.v source to its own dedicated src/ folder.   This prevents including iob_nativebridgeif.v sources when this   peripheral is not used. - Modify generation of REGFILEIF inverted register drivers to allow   multiple calls to the main() function without breaking the registers.

+++ iob_regfileif_setup.py
@@ -5,6 +5,7 @@ sys.path.insert(0, os.getcwd()+'/submodules/LIB/scripts')
 import setup
 import mkregs
 import shutil
+import copy
 
 name='iob_regfileif'
 version='V0.10'
@@ -54,7 +55,8 @@ def main():
     setup.setup(sys.modules[__name__])
 
     #### Invert registers type to create drivers for Secondary system
-    for table in regs: 
+    inverted_regs = copy.deepcopy(regs)
+    for table in inverted_regs: 
         for reg in table['regs']:
             if reg['type'] == 'W': reg['type']='R'
             else: reg['type']='W'
@@ -63,7 +65,7 @@ def main():
     mkregs_obj = mkregs.mkregs()
     mkregs_obj.config = confs
     # Get register table
-    reg_table = mkregs_obj.get_reg_table(regs, False)
+    reg_table = mkregs_obj.get_reg_table(inverted_regs, False)
     # Create inverted register hardware
     mkregs_obj.write_hwheader(reg_table, build_dir+'/hardware/src', f"{name}_inverted")
     mkregs_obj.write_hwcode(reg_table, build_dir+'/hardware/src', f"{name}_inverted")
@@ -107,9 +109,9 @@ def main():
     with open(f"{build_dir}/hardware/src/{name}_inverted_swreg_def.vh", "w") as file: file.writelines(lines)
 
     #### Create params, inst_params and conf files for inverted hardware. (Use symlinks to save disk space and highlight they are equal)
-    os.symlink(f"{name}_conf.vh", f"{build_dir}/hardware/src/{name}_inverted_conf.vh")
-    os.symlink(f"{name}_params.vh", f"{build_dir}/hardware/src/{name}_inverted_params.vh")
-    os.symlink(f"{name}_inst_params.vh", f"{build_dir}/hardware/src/{name}_inverted_inst_params.vh")
+    if not os.path.isfile(f"{build_dir}/hardware/src/{name}_inverted_conf.vh"): os.symlink(f"{name}_conf.vh", f"{build_dir}/hardware/src/{name}_inverted_conf.vh")
+    if not os.path.isfile(f"{build_dir}/hardware/src/{name}_inverted_params.vh"): os.symlink(f"{name}_params.vh", f"{build_dir}/hardware/src/{name}_inverted_params.vh")
+    if not os.path.isfile(f"{build_dir}/hardware/src/{name}_inverted_inst_params.vh"): os.symlink(f"{name}_inst_params.vh", f"{build_dir}/hardware/src/{name}_inverted_inst_params.vh")
 
     #### Create inverted register software
     mkregs_obj.write_swheader(reg_table, build_dir+'/software/esrc', f"{name}_inverted")
